content-worker/requirements.txt
# Temporal & Workflow
temporalio>=1.7.0

# Google/gRPC dependencies - pin compatible versions to avoid resolution conflicts
grpcio>=1.62.0,<2.0.0
grpcio-status>=1.62.0,<2.0.0
proto-plus>=1.24.0,<2.0.0
protobuf>=4.25.0,<6.0.0
googleapis-common-protos>=1.63.0,<2.0.0

# AI/ML & Generation - use slim version with only needed providers
pydantic-ai-slim[openai,anthropic,google]==0.1.4
google-generativeai>=0.8.0
openai>=1.0.0
anthropic>=0.40.0

# Let pydantic-ai-slim control transitive dependencies (logfire, opentelemetry, etc.)

# Database
psycopg[binary]>=3.1.0

# Memory & Knowledge Graph
zep-cloud>=2.0.0

# Image Generation & Processing
replicate>=0.25.0
cloudinary>=1.40.0
Pillow>=10.0.0

# Web Scraping & Research
firecrawl-py>=0.0.5
exa-py>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0

# Search APIs
google-search-results>=2.4.2
linkup-sdk>=0.1.0

# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0

# Dashboard UI
streamlit==1.31.0

# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0
ulid-py>=1.1.0
slugify>=0.0.1
python-slugify>=8.0.0

# URL handling
urllib3>=2.0.0
validators>=0.22.0
aiohttp>=3.9.0

# Video processing
mux-python>=5.0.0

# Force rebuild - 2024-11-22
